backend/pyproject.toml
[project]
name = "backend-assignment"
version = "1.0.0"
description = "FastAPI backend with JWT auth, RBAC, and MongoDB"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "motor==3.4.0",
    "pymongo==4.7.2",
    "python-dotenv==1.0.1",
    "pydantic[email]==2.7.1",
    "pydantic-settings==2.2.1",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "httpx==0.27.0",
    "slowapi==0.1.9",
    "mongomock-motor>=0.0.36",
    "bcrypt>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
package = false
